refactor(crypto): rename db field encryption modules [#411]
