+++ hushline/crypto/__init__.py
@@ -2,7 +2,7 @@
 A subpackage to organize hushline's cryptographic functionalities.
 """
 
-from .db_row_encryption import decrypt_field, encrypt_field
+from .db_field_encryption import decrypt_field, encrypt_field
 from .pgp_tools import encrypt_message, is_valid_pgp_key
 from .secrets_manager import SecretsManager
